Merging initial prototype development into master branch
1. Adjustments to get the application deployed and working on Heroku:

Changed build, Proc, prop files to be Heroku friendly (fat jar w/port variable)
Created a config class to make the database URL work properly on Heroku

Note: the DBUnit tests started failing here.

2. Built out the principal USSD menus in AATApiTestController

Created home page, first user epic (send a message), and some basic profile
menus working, such as renaming and so forth.

3. Added Lombok, and Many-to-Many relatinoships

Created relationships in model between groups and users who are their
members. Linked that to the join table. Added logic in the controller
to create a group with the numbers that a user inputs in the second
step of the meeting organizer. Next: Provide a way to name the group
prior to sending out a message (or, when next log in). Also added Lombok to
Maven dependencies and used for  hashcode, equals and to-string implementations.

4. Wired up the controller and menus to persistence

Made the controller create a user instance if one didn't exist already, based
on incoming phone number, and to create a group with members once numbers are
added by the user. Created some basic string processing stubs, for turning a
user inputted phone number into the database format, etc. Also, implemented
the UserRepository interface by declaring findByPhoneNumber method. This last
also broke the DBUnit tests (independent of break above).

5. Created messaging method to send to AAT SMS gateway

Built the final stage of the first user epic, sending out the message
to the group.

6. Started group persistence and management epic.

Built the group management menu, in part, and added methods to check a user's
group memberships and provide a menu of them before sending message. Created
basic methods and options for renaming user and group, menus to do that, and
methods to handle users and groups without names. Still need to prompt to name
oneself upfront and name group upfront. Also created menus and methods to
unsubscribe from a group, add a member, and delete a group. The latter are
in desperate need of permissions logic and validation, etc.

7. Started stubbing out the web application

Built a first couple of forms to stub out the web application that will
allow those with a desktop to set up groups, act on them, etc. The first
such views and controller methods ask for a phone number, call up the
user, and then have some fall backs if the user doesn't exist or hasn't set
a display name. Lots of obvious to-dos range from creating actual security
to doing the stylesheets and so on. Hence for the moment just a stub.