Added Lombok, and Many-to-Many relatinoships
As a minor adjustmet, added Lombok to the Maven dependencies and
ticked off the to-do on using it for the hashcode, equals and tostring
implementations. Also tried to wire up the many-to-many relatinoship
among groups and users, and wrote an extra test on the group CRUD
repository for it. Unfortunately, causes tests to break, to do with
tables in test database seemingly not being cleared, not clear why.