Linked up to the repository, and trying out free text entry
Added in a check on the first page if user exists, and, if not, auto
store them in the repository. Also added checks in subsequent menus
to pick up latent errors. Made the second menu in the meeting organizer
a free text entry and then added the check to see what we get back.