Added many-to-many relationships and invoked in group creation
Created relationships in model between groups and users who are their
members. Linked that to the join table. Added logic in the controller
to create a group with the numbers that a user inputs in the second
step of the meeting organizer. Next: Provide a way to name the group
prior to sending out a message (or, when next log in).