Created messaging routine to send to AAT SMS gateway
Built the final stage of the first user epic, sending out the message
to the group. Still need to get our username and password for the SMS
gateway, however. Also, need to capture and process the response from
the gateway--is an XML file much like the USSD spec. To set up.