Started group management menu (and added spec files)
Began to build the group management menu, by calling the user's
groups and generating a menu that will pass groupIds to the next
menu. For unnamed groups, uses the timestamp to avoid blank option
lines. Next to do is to build logic to ask for a group name when
user starts, and probably also to capture a display name for the
user themselves. Also a lot of validation needed, and probably
refactoring for speed and robustness. Also added the USSD and SMS
gateway specs as PDFs into a lib folder (may be wrong location,
others should feel free to move).