Give user list of groups to select when calling a meeting
Restructured the meeting flow to insert story: as a user who has created
a group in the past, I can see a list of my groups before I need to enter
numbers so that I can reuse work I've done in the past. Also stubbed out
the user profile menu. Next to do: setting a group name and a user profile
name, so that users don't need to keep remembering these.