Started stubbing out the web application
Built a first couple of forms to stub out the web application that will
allow those with a desktop to set up groups, act on them, etc. The first
such views and controller methods ask for a phone number, call up the
user, and then have some fall backs if the user doesn't exist or hasn't set
a display name. Lots of obvious to-dos range from creating actual security
to doing the stylesheets and so on.