Filled out some functions and features for the presentation demo
For USSD demo, added the "From <user>" to the front of messages that are sent
out. For the web demo, finished the critical forms that allow the creation of
a new group and the bulk renaming of members of a group once it's created.
Only things left to do (for demo) are a bit of tidying up of web menus and
making them display right-sized on a mobile. Note: Given time constraints,
have created a couple of methods that clearly belong in service layer instead
of the controller/view, including the cardinal sin of copying across a method.
These are last such additions/sins.