Added some final features and stubs for demo presentation
Finalized a few features and refined others to get ready for demo presentation
on August 2nd. Included: greeting user by name, if name is set, a form to create
a new group with users and names, and various adjustments in the view templates,
such as back to home buttons, and so forth. Also, more methods that need to
move into a service layer, such as getting a list of groups for a user, and
so forth.