Changed the USSD opening menu for length, displayname
MTN was not working because it restricts opening menu to 140 characters and
ours was 148. Also, created a logic to check if the user is unnamed and this
is not one of their first uses, and, if so, ask for a name. Also factored out
the welcome menu so it can be called in multiple places.