Integrated demo features into module refactor
Consolidated a few additional methods in web app (and small change in model) to
align demo with the newly refactored module structure. Several of the methods
in the Controllers and the User class should probably move to the service layer.
Also moved around some of the template files. To do is convert the old mustache
templates into thymeleaf and wire all together with Spring Security.