Expanding USSD test suite
Added several extra tests to the webapp module, to do integrated tests on the
USSD controller. Tests now check for user save/create, user rename and use of
name on subsequent access, as well as creating a group from a list of phone
numbers. The last test is generating an error on the role table, for some
reason. Need to fix that, as well as understand how the integrated database is
storing / dropping rows, and will then add more asserts and a few more tests.