Extended and improved USSD tests, using new group services
Improved the USSD controller tests, now using the methods to find last created
group. In doing so, added a couple of methods to the service layer to give
some separation to repository, and for naming simplicity. Also did some minor
tinkering with string handling and started to stub out USSD messages, for
localization of front end.