Continued expanding tests, menu refactoring, checking length
Added methods to USSDMenu to compute the length of the return menu, and added
a method in the controller to check that a menu will be shorter than the
aggregator-prescribed length. Used that for a bit of checking on the longer
menus, and trimming down. Likely want to build better error/exception handling
on that, and come up with good ways to truncate, rather than having to manually
recode. Also expanded the home-screen test to check that the first USSD menu
is exactly as it should be. Finally, continued refactoring the path mappings
etc to gradually clean out the literals.