Continued abstraction of USSD menus
Adjusted the construction of USSD menu to strip out the object referring to its
own URL, which was causing redundancy. Now the keys all serve as URL endings,
with standard constants for the sub-menus. Also added a method to calculate
the size of a returned menu (in USSDMenu), and a method in the controller to
check if the menu is too long for USSD. Using that, trimmed some parts of text.
Next: make sure this isn't breaking anything on the demo, and then extend across
all USSD controllers. Then, work out a way to do pagination properly.