Set up language selection for USSD menus
Created basic structures and methods for implementing i18n in USSD menus. Since
we don't have a session context from browser to set language, this has to be
a user profile attribute set in the domain. Then we use a messageSource and
default to English if the language isn't set. Put in some (badly translated)
Zulu messages, to get it started. Still need to do a bunch of refactoring as
well as switch over all the other menus to using the external properties file,
even before we are finished with translations.