i18n structures implemented throughout USSD controllers
Completed replacement of literal string coding for prompt and option text in
USSD controllers (with a couple of exceptions for system/error messages). Now
ready for local languages--have set up a dummy file for Zulu, with a couple of
strings. Next is to fix the JPA error coming up on the deployment platform plus
add the actual translations. Along the way did some refactoring and general
tidying up, as well as slight tweak to tests.