Implemented initial Event model and services in USSD meeting call
Brought the Event model and services into the USSD meeting controller. Used it
to make the menus much more flexible and abstract -- instead of passing literals
around, they pass the eventId, the name of the parameter they are changing, and
the value. Should allow easy adjustment of flow. Also starts to bring events
into the system. Along the way stubbed out some methods for the event services.
One thing to note is, since not sure yet of how strict we are going to be on
user inputted day/time, did not convert to timestamp as storage, but kept as
two strings. Probably something to address in future.

Last note: there are still a couple of tests that should pass but aren't, in
which an event that has been modified is returning null date & time. Not clear
why as same sequence of calls clearly results in non-null values returned in
production environment. Leads to flagging of issue of generally cleaning up
null / non-null definitions in the data model.