Fixing/improving group creation in USSD controller and services
Built out a group creation menu set in the group management section that asks
the user repeatedly for more numbers until they press '0'. Creates and expands
the group accordingly, then names it. Along the way, changed the service layer
to expect a list of phone numbers rather than do the number splitting in there.
The only interface that will require such splitting is the USSD interface, on
the others we can and should have separate text boxes and form validation.

Next to do are to implement this sequence in the meeting caller, as well as
build the number splitting & validation into the super class (USSDController).
Also, along the way, fixed up the feature to prompt for a group name if the
user has one that's unnamed.