Improved group addition menus, added subject to meeting menu
Created a "create new group" menu under group section of USSD menus, and used
it to implement improved phone number error handling and group creation. Now
the controller class will sort input into valid and invalid phone numbers, and
use the valid ones to create a new group, while asking the user to correct the
invalid ones (or just skip). Also, created a loop so user can enter numbers one
by one. Extended this to the meeting organizer tool, so that also does a loop
asking for numbers in batches and splits them into error and valid numbers.
Along the way did a general clean up of the logic in that controller, as well
as adding a prompt for a subject. Still need to do some refactoring, and, in
particular, timestamp processing so the meeting organizer integrates into the
eventlog service and can do reminders.