Improved group addition menus, added subject to meeting menu
Created a "create new group" menu under group section of USSD menus, and used
it to implement improved phone number error handling and group creation. Now
the controller class will sort input into valid and invalid phone numbers, and
use the valid ones to create a new group, while asking the user to correct the
invalid ones (or just skip). Also, created a loop so user can enter numbers one
by one. Last, added a subject menu to the meeting flow, which sets the name
of the event, and stubbed a menu to ask for that subject.