Messy start to extending new group functionality to meeting org
Starting adding functionality for repeat number entry in the group creation
sub-flow of meeting organizer. However, realized that some deeper surgery was
needed on the meeting org group creation methods, so this is a bit of a mess
at present and is failing tests. Will do a proper fix and full refactor and
then push a clean and working version to master.