Minor tweaks and fixes to USSD menus and users
Some house cleaning on the USSD menus, in preparation for next epics, namely
short code processing, date and time parsing, and building out language support.
Specifically: fixed language selection so it works (Spring was not picking up
the za_zu file, and gain is infinitesimal in adding country characters); also
added a couple of lines to web user preset to make username phone number if
username is null (deals with a legacy issue, but may remove after some later
data cleaning); and then put in place some basic methods to recognize when the
standard USSD string is followed by some other digits and break those up into a
list. Next to do is to write the methods to process that list.