Fixed up basic skeleton for group token join
Got the persistence and mapping of GroupTokens to work with groups, and then
wrote the controller menus to generate a token. Also put in skeleton methods
to take a token and add the user to the group matched to that token. Still
need to do quite a bit of validation as well as general user flow smoothing.