Began refactoring of web app
Moved around a bunch of the template pages so that the folder structure was
a bit cleaner and simpler to follow, and rewired the controllers appopriately.
Also removed the event listing on the home page, as it was both throwing an
error for any user that had a group and on the most likely UX/UI, uneccessary.
Last, started a small bit of fiddling around on templates. Next to do are to
extend the group creation pages. Also may revisit some of the CSS and general
theme, which is a bit busy -- general feel should be stripped down & simple.