i18n for meeting notice template, fixing a range of bugs in menus
Switched the meeting notification from assembling a notice in the text to
externalized via i18n. Now calls a template, which can also be used in other
parts of system. Besides that, fixed a range of errors and minor message bugs
that have appeared over the last few days in demos.