Adding date parsing, plus a range of other minor changes
Imported a utility library for parsing user input to generate timestamps, and
implemented in USSD controllers. Also cleaned up the Event model so it only
needs one string for day and time (separation was from deprecated menus). So
far parsing seems to work, but needs more testing. Also base URL and code length
in USSD controllers to env variables, added a few Zulu messages, and switched
GroupRepository to implement Paging, for next to do of paging.