Added group pagination, and range of fixes and adjustments
Added pagination to the Group repository, so that we avoid having overly long
USSD menus. Started to use pagination in the menus, but has a ways to go. Also
fixed the issue with group creation on the Web app, by introducing a lazyList,
and changed timnestamp/localdatetime to see if that fixes token issue. Else,
added some Zulu menus, put the SMS username & password into the messaging
layer, and poked around the messaging layer.