Token creation and sub-groups via web application
Did the (very) basic wiring for generating a group joining token via the web
application, as well as creating a subgroup. Along the way, also fixed some
bugs in service layer, to prevent duplicate user adding and to properly respond
and add to group when user inputs token code in USSD string.