Further build our of web application
In the web application, built out pages and views to create subgroups and add
them, as well as to create a token. Also fixed the meeting call to use a date
time mask, and convert into a timestamp. Since we need more information in the
group create form than is in the Group entity, and Thymeleaf / Spring MVC don't
handle that well, also built a wrapper class in the web app model. Can use that
for a property editor/convertor later, if there's a need. Has temporarily caused
a break in some of the forms, which will fix tomorrow.

Also added quite a few more Zulu translations of the USSD menus, plus the
jQuey utility for masking the date/time input.