Built token & subgroup pages for web application
Created pages to close and extend group tokens, as well as to generate them
on group creation. Also added option in meeting send to notify all members
of subgroup or not. Made a few minor changes in controllers and in services
to cater for these. Next is to do proper token code generation, to be able
to view the subgroup memebrs, etc. etc.