Range of new features, somewhat buggy.
Working on a range of front end improvements, and some new features, but still
need to make all click, and test properly. Trying to get i18n to work on
services layer, without success. Group linking to parent done, and free form
messages, but both of them a little buggy.

Also, refactored entities so User repository just returns a single User on
findByPhoneNumber, since the latter is unique, getting the iterator was
redundant.