Added 'unsubscribe' to web app, fixed parent linking, more Zulu
Added an option to the web app to unsubscribe from a group, so that people can
trim down the number of groups they belong to, along with a confirmation page.
Also fixed some issues with the page to link a group to a parent, though still
need to use new service methods to control for infinite loops. Last, added some
additional Zulu phrases.