Implemented RSVP functionality in web application
Connected the web application to RSVP functionality, which included a few tweaks
on services and core layer, e.g., methods to set yes/no simply, to obtain a map
of responses to date, send out a different meeting notice, and so forth. Next to
do is to start integrating in the USSD layer. Some of the changes and fixes will
probably need refactoring  in future to make efficient, etc.