Implemented RSVP functions in USSD
Added check to USSD home controller if user has any outstanding events to RSVP
and if they do, prompts them to RSVP. Records the response and then gives a
home/exit option. Also added/tweaked some helper functions, and created a new
Enum to make it clearer what's happening on the start menu when we have things
we need user response for. All of the above needs a lot of testing before being
promoted to production. Bones in place for votes, though.