Implemented functions to resume if cut-off during USSD
Added methods in meeting creation methods, in USSD controllers, to store the
user's location in the USSD menu, and then be able to jump back to it if the
session times out. Needs a lot more testing but seems to work at present. Also
removed the services-bypass from webapp in one of the methods, need to go through
rest and remove in full.