Some fixes to front end, and added a step in modifying meetings
Fixed a couple of bugs on the front end related to responding to RSVPs, and
added a new step in viewing/modifying upcoming meetings. Also, moved a repeated
step -- pulling fields from an event to populate a message -- into services
layer, as is called in multiple places and uses a bunch of getters, so better
in services than front.