Stubbed out next wave--admin web app, meeting relays
(Incomplete) Began to build an admin suite for the application, via the web app,
which will enable us to manage the system once users start signing up and
permissions are in place. Also, added a field to the event entity to flag whether
it should be open to "relaying", i.e., forwarding. Use is in the future, but
decided to add now. Finally, changed group controller to only call services,
which required the addition of some service level methods.