Added flag to User to distinguish if has initiated session
Based on use so far, added a flag to the User object to distinguish between
those which have been entered by another user via a group creation, and those
who have actively initiated a session themselves. Needed both for analytics of
our own, plus for use case of setting language -- have found some users don't
realize language option exists but is buried. This brings it to the front of
the platform. Then also added a range of tweaks to web application and USSD
interface.