Implementing group default language functions
Added methods and front end for "group default language" functionality, which
allows setting a language and having new-to-the-system users who are added to
it given that language as default (but they can still select and override on
their first system use). Also fixed a couple of minor bugs, and added a needed
line to Travis config file.