Started wiring up accounts, and fixed a few bugs
Began to create the account and paid group services and stubbed out some methods
for their front end. All of it needs thorough testing, which have not built yet,
before being exposed even to pilot users. Also did a few fixes on the USSD and
web interfaces for bugs, specifically closing #19 and #16.