Initial implementation of USSD logbook menus
Creation of menu flow and basic functions for USSD menus that implement the
logbook entity and services. Still need to add i18n messages (as written will
throw errors outside tests), add tests, and verify that works properly in
general. I.e., untested and not ready for production but checking in just in
case others want to see/inspect, including for possible changes to entities
based on this (e.g., referencing #52 and #53).