Restored user rename, group rename, finished welcome msgs.
Restored the USSD prompts to rename users and to ask for names for unnamed
groups, based on observing user behaviour to date -- with lots of unnamed users
and fragmentary groups. If group has never had a complete event called, also
allow it to be made inactive. All of the above needs to be optimized heavily,
as involve quite a few DB calls. Also, added in the welcome messages. Closes
issue #43, closes issue #39, closes issue #36.