Added recorded field to logbook, plus other changes, closes #53
Created a field 'recorded' that is used to make sure we don't view, record or
use half-completed logbook entries from the USSD menu. Added scripts to create
it. Also added a field to paid accounts for extra logbook entries, which will
use shortly. Last, added a lot of stubs to the messages files, as the logbook
menus were missing, plus many others. Does a little for #31 but much left.