Building out account functionality, roles, permissions
Range of changes to start making progress on accounts, paid groups, roles and
permissions. Implemented a form that allows sys admin to adjust roles of group
members (basis for later form for any group org). Also created AccountController
to enable functions for an account administrator to designate groups. At present
limited to only ones they are part of (need to change that). In implementing the
removal of groups, added a field (archived account ID) so we can still find the
relationship, and inserted a dummy account to set the current account to zero.
Along the way fixed a few other minor issues, and created a proper 403 error page.