View event history for paid groups via accounts
Built out methods and pages to view lists of events for a group in the past.
This gets further to completing that functionality. Left is to integrate costing
as well as allow someone to find groups. Also, adjusted the base controller so
we don't ping DB for user all the time. Closes issue #32, GRUSSDWV1-38.