Finished view group history for paid accounts
Completed page to view records of events and meetings for a paid group. Can
extend next to work for standard groups, as added an EventWrapper that is multi
purpose (or should be). Last things are to build ways to find groups to add them
to the paid account (and hence more generally), and to view the details of the
prior record. Thereafter, finishing Web Phase I is about building out all the
tests so we have complete code coverage and a range of UI fixes.