Initial means to find groups not part of, also logbook fixes
Added a field to groups, 'discoverable', set to false by default, which if true
allows the group to be found through a search on name. Built that search into
a crude form for account admin (which also allows searching by token). Aside
from that, did a minor refactor in group services, and then a few tests and fixes
in the USSD LogBook controller. Finally, a little closer to getting the group
permissions to work.

GRUSSDWV1-81 #comment Done for account admin
GRUSSDWV1-70 #comment Done bar more testing