GR-USSDWV1-119, 103, 125, 126
Principal change is integration of role setting in all group creation and member
addition methods. All can be toggled through a boolean added to the methods. Also
refactored GroupManager almost entirely, stripping out a large amount of redundancy
and poor methods, as well as passing the user doing the action around so it gets
recorded in the logs properly. To get roles to work with USSD (and tests) had to
also pass the user making the change to access control. All of these methods need
to become async quickly, once async no longer breaks bean injection.