GRUSSDWV1-128
Created a new cache (userUSSDMenu) to store current position, and retrieve it.
Allows us to avoid DB pings at the beginning of the USSD section. Consider a
way to also cache the interruption menu strings (or language code) so that the
first menu on resume can bypass DB altogether -- if user is timing out, on
resume want to be as fast as possible.