Cleaned up keystores, took out Joda, scaffolded role display
Removed redundant keystores and switched over those in webapp and test to make
clear they're only for integration testing and local pg (i.e., do not apply on
any production environment). Also removed Joda from the pom and replaced it with
Java 8 localdatetime in the one place that was using it. Last, added a quick page
to display the roles in a group and their permissions (not linked to main group
view yet, mostly for last debugging and checking on legacy groups).